Commit 19/10 conexión ruta y fetch para vista catalogo
