+++ src/front/js/component/elementoCatalogo.jsx
@@ -2,13 +2,7 @@ import React, { Component } from "react";
 // import { useContext } from "react";
 import { Link } from "react-router-dom";
 
-export const ElementoCatalogo = ({
-  ubicacion,
-  rating,
-  amenidades,
-  reviews,
-  descripcion,
-}) => {
+export const ElementoCatalogo = ({ name, description, images }) => {
   // const { store, actions } = useContext(Context); //consumir el contexto
   return (
     <div>
@@ -24,30 +18,24 @@ export const ElementoCatalogo = ({
           <div className="card-body">
             <h1 className="card-title">Titulo de la casa</h1>
             <h6 className="card-text">
-              <h5 className="titulo-catalogo">Esta es la ubicación</h5>
-              {ubicacion}
+              <h5 className="titulo-catalogo">Esta es el nombre</h5>
+              {name}
             </h6>
-            <h6 className="card-text">
-              <h5 className="titulo-catalogo">
-                Este es el rating que tiene la propiedad
-              </h5>
-              {rating}
-            </h6>
-            <h6 className="card-text">
+            {/* <h6 className="card-text">
               <h5 className="titulo-catalogo">Las amenidades son</h5>
               {amenidades}
             </h6>
             <h6 className="card-text">
               <h5 className="titulo-catalogo">Estos reviews tiene el sitio</h5>
               {reviews}
-            </h6>
-            {/* <p className="card-text">
-                <h1>Fotos del sitio</h1>
-                {item.fotos}
-              </p> */}
+            </h6> */}
+            <p className="card-text">
+              <h1>Fotos del sitio</h1>
+              {images}
+            </p>
             <h6 className="card-text">
               <h5 className="titulo-catalogo">Descripción</h5>
-              {descripcion}
+              {description}
             </h6>
           </div>
         </div>
