+++ src/front/js/layout.jsx
@@ -2,14 +2,14 @@ import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ScrollToTop from "./component/scrollToTop";
 
-import { Home } from "./pages/home.js";
+import { Home } from "./pages/home.jsx";
 import { Login } from "./pages/login.jsx";
 import { DetailView } from "./pages/detailView.jsx";
-import { Demo } from "./pages/demo";
-import { Single } from "./pages/single";
+import { Demo } from "./pages/demo.jsx";
+import { Single } from "./pages/single.jsx";
 import { Catalogo } from "./pages/catalogo.jsx";
 import { Postspace } from "./pages/postspace.jsx";
-import injectContext from "./store/appContext";
+import injectContext from "./store/appContext.jsx";
 
 import Navbar from "./component/navbar.jsx";
 import { Footer } from "./component/footer";
