+++ src/front/js/pages/postspace.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { Context } from "../store/appContext";
+import { Context } from "../store/appContext.jsx";
 import "../../styles/home.css";
 // import { Container, FormGroup, Input } from "reactstrap";
