+++ src/front/js/pages/single.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from "react";
 import PropTypes from "prop-types";
 import { Link, useParams } from "react-router-dom";
-import { Context } from "../store/appContext";
+import { Context } from "../store/appContext.jsx";
 import rigoImageUrl from "../../img/rigo-baby.jpg";
 
 export const Single = (props) => {
