+++ src/front/js/store/appContext.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import getState from "./flux.js";
+import getState from "./flux.jsx";
 
 // Don't change, here is where we initialize our context, by default it's just going to be null.
 export const Context = React.createContext(null);
@@ -24,6 +24,7 @@ const injectContext = (PassedComponent) => {
     );
 
     useEffect(() => {
+      state.actions.traerCatalogo();
       /**
        * EDIT THIS!
        * This function is the equivalent to "window.onLoad", it only runs once on the entire application lifetime
